Update README.md
Minor clarifications in the readme.

When i set this up I was under the impression this would impersonate only rather than act as True man in the middle. So i was a bit surprised to find this was performing a DNS query and then  still talking to the ecosense servers.